refactor: split vp level partition and vp tree
